+++ lang/en/src/Export/ExportAction.php
@@ -1,0 +1,6 @@
+﻿<?php
+$MESS['MB_ADMIN_KIT_EXPORT_PERMISSION_DENIED'] = 'Export permission denied.';
+$MESS['MB_ADMIN_KIT_EXPORT_ACTION_NOT_ALLOWED'] = 'Export action is not allowed.';
+$MESS['MB_ADMIN_KIT_EXPORT_ALL_DISABLED'] = 'Exporting all records is disabled by default. Select records or pass an explicit filter.';
+$MESS['MB_ADMIN_KIT_EXPORT_FILTER_DISABLED'] = 'Export by filter is disabled for this action.';
+$MESS['MB_ADMIN_KIT_EXPORT_UNSUPPORTED_FORMAT'] = 'Unsupported export format.';