+++ lang/ru/src/Export/ExportAction.php
@@ -1,0 +1,6 @@
+﻿<?php
+$MESS['MB_ADMIN_KIT_EXPORT_PERMISSION_DENIED'] = 'Нет прав на экспорт.';
+$MESS['MB_ADMIN_KIT_EXPORT_ACTION_NOT_ALLOWED'] = 'Действие экспорта недоступно.';
+$MESS['MB_ADMIN_KIT_EXPORT_ALL_DISABLED'] = 'Экспорт всех записей отключен по умолчанию. Задайте явный фильтр.';
+$MESS['MB_ADMIN_KIT_EXPORT_FILTER_DISABLED'] = 'Экспорт по фильтру отключен для этого действия.';
+$MESS['MB_ADMIN_KIT_EXPORT_UNSUPPORTED_FORMAT'] = 'Неподдерживаемый формат экспорта.';